Made functionality to login
Made validation of user validation of login route with user_class. If user send form instance of user_class is created and then by function verify_user and then it is checked whether a user with such a password and login exists, if so the session variable that contains the user's login is created. That allows user to be redirected to index.html